backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.2.0
psycopg2-binary==2.9.11
pytz==2026.1.post1
sqlparse==0.5.5

dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
cloudinary==1.39.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
